Phase 2: i16 quantization (scale 5000) — lossless detection, 2 instances
Migrate the index + runtime from uint8 to int16 (scale 5000), on top of the
Phase 1 KD-tree. Removes the uint8 quantization floor that bounded detection.

Why i16 / scale 5000 (measured, leave-one-out over the real 3M dataset vs the
float oracle that defines the test labels):
  - exact uint8 5-NN: 0.55% decision error (the floor Phase 1 could not cross).
  - exact i16 (scale 5000 AND 10000): 0 errors / 6000 — LOSSLESS for the
    deny/approve decision. Scale 5000 chosen: worst dist_sq = 5e8 << INT32_MAX
    (4x margin), vs scale 10000's thin 147M margin.
  => detection_score ~735 -> ~3000 (+~2200), now that Phase 1 made search exact.

Format v13: vectors are 16 int16 lanes (14 dims + 2 zero pad) = 32 bytes = one
ymm; sentinel (null last_transaction, float -1) -> -5000. KD node AABBs are i16.

AVX2 kernels rewritten (distance.s, distance_batch.s): VPSUBW + VPMADDWD (no
min/max/zero-extend). bound_dist_sq uses VPSUBW + VPMAXSW(edge) + clamp-to-0.
Overflow-safe: per-lane diff <=10000 (fits i16), diff^2 <=1e8, vpmaddwd pair
<=2e8/i32 lane, 8-lane sum <=5e8 < INT32_MAX.

Topology 4->2 instances (i16 index is 118MB, no longer fits 4x80MB):
  2 API x (0.45 CPU / 160M) + haproxy (0.10 / 30M) = 1.0 CPU / 350MB.
  Each API holds the full index. (c-impl/docker-compose.yml was already 2x.)

Validated locally (all green) on the real 3M v13 index:
  - 269,491 nodes, maxDepth 32, index 118.4MB.
  - make test: distance(i16 kernels), vectorize(i16), json, search all pass.
  - EXACT recall: 0 dist-set & 0 fraud-count mismatches vs brute force over
    1000 queries (tests/test_kd_exact.c).
  - 6.76us/query AVX2 single-thread (1.5x Phase 1 uint8; still ~28x vs old scan).
  - Container under -m 160m --cpus 0.45: boots, serves, OOMKilled=false,
    ~1.6MiB anon RSS (118MB index is reclaimable file cache).
  - Full stack via HAProxy :9999: 500/500 HTTP 200, fraud->false/1.0,
    legit->true/0.0.

Projected: detection_score 735 -> ~3000; p99 stays low (KD + i16 6.76us/query).
Combined with Phase 1: final ~1898 -> ~5000.

Files: src/{compat.h,distance.{h,s},distance_batch.s,vectorize.{h,c},index.{h,c},
search.{h,c},server.c} i16; tools/build_kdtree.go v13; docker-compose.yml +
haproxy.cfg 2 instances; tests/* updated for i16.

+++ c-impl/src/distance.h
@@ -3,30 +3,26 @@
 
 #include "compat.h"
 
-/* Squared Euclidean distance between two 16-byte uint8 vectors.
- * Bytes 14 and 15 of both `q` and `ref` MUST be 0 (padding contract);
- * they participate in the sum.
- *
- * Implementation: AVX2 in distance.s (Haswell). Matches the Go runtime's
- * distanceSqAvx in distance_amd64.s byte-for-byte. */
-extern uint32_t distance_sq(const uint8_t *q, const uint8_t *ref);
+/* Squared Euclidean distance between two 16-lane int16 vectors (32 bytes).
+ * Lanes 14 and 15 of both `q` and `ref` MUST be 0 (padding contract); they
+ * contribute 0 to the sum. Values are i16 in [-5000, 5000]; result fits u32
+ * (max 5e8). Implementation: AVX2 in distance.s (vpsubw + vpmaddwd). */
+extern uint32_t distance_sq(const int16_t *q, const int16_t *ref);
 
-/* Squared lower-bound Euclidean distance from `q` to a partition's
- * axis-aligned bounding box (min, max, each 16-byte uint8).
- * For each dim:
- *   q < min  → (min - q)^2
- *   q > max  → (q - max)^2
- *   else     → 0
- *
- * AVX2: uses VPSUBUSB + VPMAXUB to compute edge distance per byte. */
-extern uint32_t bound_dist_sq(const uint8_t *q,
-                              const uint8_t *min_box,
-                              const uint8_t *max_box);
+/* Squared lower-bound Euclidean distance from `q` to an axis-aligned bounding
+ * box (min, max, each 16-lane int16). Per lane:
+ *   q < min  -> (min - q)^2
+ *   q > max  -> (q - max)^2
+ *   else     -> 0
+ * AVX2: vpsubw + vpmaxsw (edge) + clamp-to-0 + vpmaddwd. */
+extern uint32_t bound_dist_sq(const int16_t *q,
+                              const int16_t *min_box,
+                              const int16_t *max_box);
 
-/* Batched: computes distance_sq(q, base + 16*i) for i in [0,4) into out[0..3].
- * Implementation: distance_batch.s. Caller must ensure base[..63] is valid. */
-extern void distance_sq_batch4(const uint8_t *q,
-                               const uint8_t *base,
+/* Batched: distance_sq(q, base + 16*i) for i in [0,4) into out[0..3].
+ * `base` must point at 4 contiguous 32-byte records (128 bytes valid). */
+extern void distance_sq_batch4(const int16_t *q,
+                               const int16_t *base,
                                uint32_t      *out);
 
 #endif /* DISTANCE_H */
